+++ Src/Library/changelog.md
@@ -10,3 +10,4 @@
 ### FIXES
 - swagger response examples not honoring serializer settings #280
 - swagger request property xml examples not picked up for route params #287 
+- property injection not working on sub-classes #292